Singletons for gauge and treedb
